+++ CHANGELOG.md
@@ -8,12 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [0.10] - 2021-12-03
 ### Added
 - A *clean all* functionality for flex compilers (like `make`) that support
   it.  For the `make` compiler, this calls the `cleanall` target.
 
 ### Changed
 - Version bump.
+
 
 ## [0.9] - 2021-09-14
 ### Added
@@ -112,7 +114,8 @@ Major refactoring: cleaned up compilation and *package-lint* warnings.
 First major release.
 
 
-[Unreleased]: https://github.com/plandes/flex-compile/compare/v0.9...HEAD
+[Unreleased]: https://github.com/plandes/flex-compile/compare/v0.10...HEAD
+[0.10]: https://github.com/plandes/flex-compile/compare/v0.9...v0.10
 [0.9]: https://github.com/plandes/flex-compile/compare/v0.8...v0.9
 [0.8]: https://github.com/plandes/flex-compile/compare/v0.7...v0.8
 [0.7]: https://github.com/plandes/flex-compile/compare/v0.6...v0.7